[scripts] adapt to lower version debian and fedora

+++ syzbridge/scripts/deploy-fedora-image.sh
@@ -150,7 +150,7 @@ function compile_fedora() {
         if [ ${code_name} == "31" ]; then
             sed -i 's/.\/process_configs.sh -i kernel %{rpmversion}/.\/process_configs.sh $OPTS kernel %{rpmversion}/' kernel.spec
         else
-            if [ -z "$(grep CONTINUEONERROR ./process_configs.sh)" ]; then
+            if [ -f "./process_configs.sh" ] && [ -z "$(grep CONTINUEONERROR ./process_configs.sh)" ]; then
                 mv process_configs.sh process_configs.sh.bk
                 mv ~/process_configs.sh ./
                 chmod +x ./process_configs.sh
